requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-dump-die==0.1.5
idna==3.4
Pillow==9.5.0
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
